Fix code review issues: UB, header deps, internal API, and cleanup
- Fix integer overflow UB in loadWakeCounterFromEEPROM by casting EEPROM.read() to uint32_t before bit-shifting
- Remove heavy includes (driver.h, TFT_eSPI.h) from headers, use forward declarations instead; add includes to .cpp files
- Move getBasicAuthHeader() and fetchChecksum() into anonymous namespace in image_api.cpp
- Remove <WiFi.h> from firmware_state.h, use int8_t instead of wl_status_t
- Add static_assert for BMP row buffer size vs SCREEN_WIDTH
- Remove unused reconnectWiFi()
- Replace EEPROM.begin(128) magic number with EEPROM_SIZE constant
- Translate Swedish strings to English in serial output, display labels, and comments

+++ storage.cpp
@@ -25,10 +25,10 @@ void saveWakeCounterToEEPROM(uint32_t counter) {
 
 uint32_t loadWakeCounterFromEEPROM() {
   uint32_t counter = 0;
-  counter = (EEPROM.read(EEPROM_WAKE_COUNTER_OFFSET) << 24) |
-            (EEPROM.read(EEPROM_WAKE_COUNTER_OFFSET + 1) << 16) |
-            (EEPROM.read(EEPROM_WAKE_COUNTER_OFFSET + 2) << 8) |
-            EEPROM.read(EEPROM_WAKE_COUNTER_OFFSET + 3);
+  counter = ((uint32_t)EEPROM.read(EEPROM_WAKE_COUNTER_OFFSET) << 24) |
+            ((uint32_t)EEPROM.read(EEPROM_WAKE_COUNTER_OFFSET + 1) << 16) |
+            ((uint32_t)EEPROM.read(EEPROM_WAKE_COUNTER_OFFSET + 2) << 8) |
+            (uint32_t)EEPROM.read(EEPROM_WAKE_COUNTER_OFFSET + 3);
   return counter;
 }
